Ignore errors from specific domains
Some domains, such as nytimes.com, don't support crawlers
(or at least not this one), and always return a 500 or
404 error unless the user is logged in and has a valid
cookie. This changeset adds an option for specifying domains
to ignore and changes the thread worker to check the
URL against a list of "problem domains" before handling
any errors.

Change-Id: I9afe074fd2d52e2688a00d9e6da1cc941848a7f7
Signed-off-by: Doug Hellmann <doug.hellmann@gmail.com>