Drop curses in favor of simpler ANSI color codes
We assume all TTYs support ANSI color except on Windows, which must also
define the ANSICON environment variable.

We otherwise don't support Windows console control sequences for color
because that's a bit out of scope for this kind of script, and I'd
prefer not to introduce a dependency on something like colorama.